test(policy): move unit tests under policy subpackages
- Move failure tests to policy/failure\n- Move routing tests to policy/routing\n- Update imports to use spec package\n\nAll tests and lint pass.
